[api/integration] Make adding an integration a one-file change
The integrations composition root interleaved per-provider machinery in
a single file: each provider needed its own *ModuleParts interface,
load* function, migrations-table constant, config-flag branch, and
database/worker wiring. Adding an integration meant editing six places
in index.ts.

Introduce a uniform IntegrationModuleParts shape and an
IntegrationProviderModule descriptor, and move each provider's loader,
adapter-wiring closures, and migrations-table name into its own file
under src/providers/. A single providerModules list is the only
registration point, and createIntegrationsContext maps the parts
generically, so index.ts no longer grows as integrations are added.

No behavior change.
