+++ libs/api/integration/core/src/index.ts
@@ -1,10 +1,6 @@
 import {dirname, resolve} from 'node:path';
 import {fileURLToPath} from 'node:url';
-import type {IntegrationConnection as CoreIntegrationConnection} from '@shipfox/api-integration-core-dto';
-import {createDebugIntegrationProvider} from '@shipfox/api-integration-debug';
-import type {ConnectGithubInstallationInput} from '@shipfox/api-integration-github';
-import type {ConnectSentryInstallationInput} from '@shipfox/api-integration-sentry';
-import type {ModuleDatabase, ModuleWorker, ShipfoxModule} from '@shipfox/node-module';
+import type {ShipfoxModule} from '@shipfox/node-module';
 import type {IntegrationProvider} from '#core/entities/provider.js';
 import {
   createIntegrationProviderRegistry,
@@ -14,23 +10,15 @@ import {
   createSourceControlIntegrationService,
   type IntegrationSourceControlService,
 } from '#core/source-control-service.js';
-import {
-  getIntegrationConnectionById,
-  updateIntegrationConnectionLifecycleStatus,
-  upsertIntegrationConnection,
-} from '#db/connections.js';
+import {getIntegrationConnectionById} from '#db/connections.js';
 import {db} from '#db/db.js';
 import {migrationsPath} from '#db/migrations.js';
 import {integrationsOutbox} from '#db/schema/outbox.js';
-import {
-  publishIntegrationEventReceived,
-  publishSourcePush,
-  recordDeliveryOnly,
-} from '#db/webhook-deliveries.js';
 import {createIntegrationRoutes} from '#presentation/routes/index.js';
+import {loadEnabledProviderModules} from '#providers/modules.js';
+import type {IntegrationModuleParts} from '#providers/types.js';
 import {createIntegrationsMaintenanceActivities} from '#temporal/activities/index.js';
 import {INTEGRATIONS_MAINTENANCE_TASK_QUEUE} from '#temporal/constants.js';
-import {config} from './config.js';
 
 const packageRoot = resolve(dirname(fileURLToPath(import.meta.url)), '..');
 const maintenanceWorkflowsPath = resolve(packageRoot, 'dist/temporal/workflows/index.js');
@@ -107,177 +95,6 @@ export interface IntegrationsContext {
   sourceControl: IntegrationSourceControlService;
 }
 
-// Stable migration-tracking table names per provider database. These must NOT
-// depend on the provider's position in the module `database` array — a
-// positional name would shift if a provider is flag-disabled and silently
-// re-run migrations against existing tables.
-const GITHUB_MIGRATIONS_TABLE = '__drizzle_migrations_integrations_github';
-const SENTRY_MIGRATIONS_TABLE = '__drizzle_migrations_integrations_sentry';
-
-interface GithubModuleParts {
-  provider: IntegrationProvider;
-  database: ModuleDatabase;
-}
-
-interface SentryModuleParts {
-  provider: IntegrationProvider;
-  database: ModuleDatabase;
-  worker: ModuleWorker;
-}
-
-async function loadGithubModuleParts(): Promise<GithubModuleParts> {
-  const {
-    createGithubIntegrationProvider,
-    getGithubInstallationByInstallationId,
-    db: githubDb,
-    migrationsPath: githubMigrationsPath,
-    upsertGithubInstallation,
-  } = await import('@shipfox/api-integration-github');
-
-  async function getExistingGithubConnection(input: {
-    installationId: string;
-  }): Promise<CoreIntegrationConnection<'github'> | undefined> {
-    const installation = await getGithubInstallationByInstallationId(input.installationId);
-    if (!installation) return undefined;
-    const connection = await getIntegrationConnectionById(installation.connectionId);
-    if (!connection) return undefined;
-    return connection as CoreIntegrationConnection<'github'>;
-  }
-
-  async function connectGithubInstallation(
-    input: ConnectGithubInstallationInput,
-  ): Promise<CoreIntegrationConnection<'github'>> {
-    return await db().transaction(async (tx) => {
-      const connection = await upsertIntegrationConnection(
-        {
-          workspaceId: input.workspaceId,
-          provider: 'github',
-          externalAccountId: input.installationId,
-          displayName: input.displayName,
-          lifecycleStatus: 'active',
-        },
-        {tx},
-      );
-
-      await upsertGithubInstallation(
-        {
-          connectionId: connection.id,
-          ...input.installation,
-        },
-        {tx},
-      );
-
-      return connection as CoreIntegrationConnection<'github'>;
-    });
-  }
-
-  return {
-    provider: createGithubIntegrationProvider({
-      getExistingGithubConnection,
-      connectGithubInstallation,
-      publishSourcePush,
-      recordDeliveryOnly,
-      getIntegrationConnectionById,
-      coreDb: db,
-    }),
-    database: {db: githubDb, migrationsPath: githubMigrationsPath},
-  };
-}
-
-async function loadSentryModuleParts(): Promise<SentryModuleParts> {
-  const {
-    createSentryIntegrationProvider,
-    createSentryMaintenanceWorker,
-    getSentryInstallationByInstallationUuid,
-    persistVerifiedUnclaimedInstallation,
-    upsertSentryInstallation,
-    db: sentryDb,
-    migrationsPath: sentryMigrationsPath,
-  } = await import('@shipfox/api-integration-sentry');
-
-  async function getConnectionById(
-    id: string,
-  ): Promise<CoreIntegrationConnection<'sentry'> | undefined> {
-    const connection = await getIntegrationConnectionById(id);
-    if (!connection) return undefined;
-    return connection as CoreIntegrationConnection<'sentry'>;
-  }
-
-  async function connectSentryInstallation(
-    input: ConnectSentryInstallationInput,
-  ): Promise<CoreIntegrationConnection<'sentry'>> {
-    return await db().transaction(async (tx) => {
-      const connection = await upsertIntegrationConnection(
-        {
-          workspaceId: input.workspaceId,
-          provider: 'sentry',
-          externalAccountId: input.installationUuid,
-          displayName: input.displayName,
-          lifecycleStatus: 'active',
-        },
-        {tx},
-      );
-
-      // Promotes the verified-unclaimed row to claimed by setting connection_id.
-      await upsertSentryInstallation(
-        {
-          connectionId: connection.id,
-          installationUuid: input.installationUuid,
-          orgSlug: input.orgSlug,
-          status: 'installed',
-          codeHash: input.codeHash,
-          installerUserId: input.installerUserId,
-        },
-        {tx},
-      );
-
-      return connection as CoreIntegrationConnection<'sentry'>;
-    });
-  }
-
-  return {
-    provider: createSentryIntegrationProvider({
-      getSentryInstallation: ({installationUuid}) =>
-        getSentryInstallationByInstallationUuid(installationUuid),
-      getConnectionById,
-      connectSentryInstallation,
-      persistVerifiedUnclaimedInstallation,
-      coreDb: db,
-      publishIntegrationEventReceived,
-      recordDeliveryOnly,
-      getIntegrationConnectionById,
-      updateConnectionLifecycleStatus: updateIntegrationConnectionLifecycleStatus,
-    }),
-    database: {db: sentryDb, migrationsPath: sentryMigrationsPath},
-    worker: createSentryMaintenanceWorker(),
-  };
-}
-
-async function createConfiguredProviders(): Promise<{
-  providers: IntegrationProvider[];
-  github: GithubModuleParts | undefined;
-  sentry: SentryModuleParts | undefined;
-  workers: ModuleWorker[];
-}> {
-  const providers: IntegrationProvider[] = [];
-  const workers: ModuleWorker[] = [];
-  if (config.INTEGRATIONS_ENABLE_DEBUG_PROVIDER) {
-    providers.push(createDebugIntegrationProvider({upsertIntegrationConnection}));
-  }
-  let github: GithubModuleParts | undefined;
-  if (config.INTEGRATIONS_ENABLE_GITHUB_PROVIDER) {
-    github = await loadGithubModuleParts();
-    providers.push(github.provider);
-  }
-  let sentry: SentryModuleParts | undefined;
-  if (config.INTEGRATIONS_ENABLE_SENTRY_PROVIDER) {
-    sentry = await loadSentryModuleParts();
-    providers.push(sentry.provider);
-    workers.push(sentry.worker);
-  }
-  return {providers, github, sentry, workers};
-}
-
 export async function createIntegrationsModule(
   options: CreateIntegrationsModuleOptions = {},
 ): Promise<ShipfoxModule> {
@@ -287,17 +104,11 @@ export async function createIntegrationsModule(
 export async function createIntegrationsContext(
   options: CreateIntegrationsModuleOptions = {},
 ): Promise<IntegrationsContext> {
-  let providers: IntegrationProvider[];
-  let github: GithubModuleParts | undefined;
-  let sentry: SentryModuleParts | undefined;
-  let providerWorkers: ModuleWorker[] = [];
-  if (options.providers) {
-    providers = options.providers;
-  } else {
-    ({providers, github, sentry, workers: providerWorkers} = await createConfiguredProviders());
-  }
+  const parts: IntegrationModuleParts[] = options.providers
+    ? options.providers.map((provider) => ({provider}))
+    : await loadEnabledProviderModules();
 
-  const registry = createIntegrationProviderRegistry(providers);
+  const registry = createIntegrationProviderRegistry(parts.map((part) => part.provider));
   const sourceControl = createSourceControlIntegrationService({
     registry,
     getIntegrationConnectionById,
@@ -307,8 +118,7 @@ export async function createIntegrationsContext(
     name: 'integrations',
     database: [
       {db, migrationsPath},
-      ...(github ? [{...github.database, migrationsTableName: GITHUB_MIGRATIONS_TABLE}] : []),
-      ...(sentry ? [{...sentry.database, migrationsTableName: SENTRY_MIGRATIONS_TABLE}] : []),
+      ...parts.flatMap((part) => (part.database ? [part.database] : [])),
     ],
     routes: createIntegrationRoutes(registry, sourceControl),
     publishers: [{name: 'integrations', table: integrationsOutbox, db}],
@@ -325,7 +135,7 @@ export async function createIntegrationsContext(
           },
         ],
       },
-      ...providerWorkers,
+      ...parts.flatMap((part) => part.workers ?? []),
     ],
   };
