+++ libs/api/integration/core/src/providers/modules.ts
@@ -1,0 +1,21 @@
+import {debugProviderModule} from '#providers/debug.js';
+import {githubProviderModule} from '#providers/github.js';
+import {sentryProviderModule} from '#providers/sentry.js';
+import type {IntegrationModuleParts} from '#providers/types.js';
+
+/**
+ * Every integration the core module can compose. To add a provider, add its
+ * file under `#providers/` and append it here — nothing else in this package
+ * needs to change. Order is significant: databases are migrated in this order,
+ * so list a provider before any that depend on its tables.
+ */
+const providerModules = [debugProviderModule, githubProviderModule, sentryProviderModule];
+
+export async function loadEnabledProviderModules(): Promise<IntegrationModuleParts[]> {
+  const parts: IntegrationModuleParts[] = [];
+  for (const module of providerModules) {
+    if (!module.enabled) continue;
+    parts.push(await module.load());
+  }
+  return parts;
+}
